[#102.5.4.4.2.2.2.3.2.4.2.5.1.3.5-GSU-UPFACTOR-001]
What is the anatomy of an "upfactor"? it happens in 3 steps:

1. move and rename files to what they "should" be, without changing the
content.  This will be a broken, transitory state, but doing it as such
allows us to track exactly what we change and why -- effectively making
less noisy diffs.

( #multipart - git-stash-untracked should be a monotlith cli.rb and not
a monolith script.  it is not standalone anymore anyway.)