[#po-misc] - improve option parser nerking
  + there was some silly hackery going on with inheritable attributes
  + option syntax is no longer a struct

  + more generally, we have un-done the "struct problem" elsewhere

(since you ask, the "Struct Problem" is the one where you paint yourself
in a corner by at first using a Struct, thinking it will be short and
elegant, only to then find that it doesn't scale, it forces you to do
strange things, and encourges you to violate design, or worse, forego
it. The same thing happened with hashes in a lot of places. They are
convenient (vital event) so we made the mistake of subclassing them
all the time when we should have just made custom clases that *have*
structs etc. That is the "Struct Problem".)

  + which led to an extensive fix-up of code everywhere that used
    the old porcelaint atttribute definer, which has seen a big
    visual cleanup in this commit.

  + try to use more of the same names for things (duplicate -> dupe,
    key? -> has, keys -> names)

  + nlp for a / an (upcomming commit will use it, not tested)

  + oh and looks like we have a big move coming up so we better leave
    this commit behind for now ..