[#tm-op] (#mv 1/1) (branchdowns & renames only) (NO CHANGE IN CONTENT)
we are doing an aggressive refactor and we know now what new names we
want to use so we are first moving the old files (some of which have OK-
to-good content) to the new locations with the new names before we
continue tearing this around. this will enable us to track obsessively
the changes on a class-by-class basis, to demonstrate what we are
actually after which is a reduction in LOC without losing
functionality.

cli/option/parser/syntax.rb - unchanged. this is more to do with more, and legacy
cli/option/documenter.rb - OK
cli/option/parser.rb - OK (this will probably go away)
cli/option/documenter.rb - OK (will shrink)
cli/option.rb - BRANCHDOWN
  CLI::Option::InstanceMethods -> cli/option/instance-methods.rb (will go away)
  CLI::Option::Option_Box -> cli/option/box.rb
  CLI::Option::Option_Metadata -> CLI::Option
  CLI::Option::Recorders::OptionParser -> cli/option/probe.rb
  CLI::Option::Option_Scanner -> cli/option/scanner.rb (will be used in future..)

%s/Option_Box/Box/gc
%s/Option::Recorders/Option::Probe/gc
%s/CLI::Option::Option_Metadata/CLI::Option/gc
%s/CLI::Option::Option_Scanner/CLI::Option::Scanner/gc
_