[ts] a mother flipping generator that .. (1348)
a mother flipping generator that generates intermediate test files.

  + host proxy (for actions) is now chained
      (so that we could use chained non-ingested services properly.)
  + "tmplo" floats up a little bit
  + "verbose" floats up a little bit

  + (fixed ironically broken test/all)

ideas spawned:

  + "tree walker" could centralize a lot of messy logic

Also: "CHOKEPOINTS":

check this crazy fun out: so the test suite is annoying to run here in my cave
alone at home because it takes like 8 minutes now for the whole suite to run.
granted, that's never an excuse not to do it right, but i believe i am finding
a multiplier effect here:

with our `divide` plugin in the test suite, we can divide the test suite
up (randomly) into some number of chunks (tests under one subproduct are
treated as a unit).

in this series of meta-tests, we ran the whole test suite four times. the
first time we ran it in four little chunks, the second time was in three
chunks, and so on. what how dramatically the cumulative time it takes to
run the whole test suite increases as you reduce the number of chunks!

ONE (run all 27 subproducts in a series): (510.48 seconds)

  num tests   duration of running that "chunk":

 1348      8 minutes 30.48 seconds

TWO divide two random chunks: 1 minute, 46.91 seconds (106.91 seconds)

  751      1 minute 23.46 seconds
  597               23.45 seconds
 1348

THREE: 37.04 seconds!!

  396        3.94 seconds
  592       20.83 seconds
  360       12.27 seconds
 1348

FOUR: 36.32 seconds (OMG)

  249       11.14 seconds
  575       20.12 seconds
  239        2.52 seconds
  285        2.54 seconds
 1348

savings conclusion:

dividing it into two ran it in 20.94% of the time ( 4.77 X faster )
dividing it into three ran it 7.25% of the time! ( 13.78 X faster!!)
dividing it into four ran it 7.11% of the time!  ( 14.05 X faster!)

Having it run 14 times faster seems signficant.

Conclusion: well, a couple of things, like three of them are some possible
corollaries of this. 1) wouldn't it be fun to fix ruby 2) maybe i am doing
it wrong at this point by .. etc  leaning on ruby so much and/or failing
to cull down the scope and breadth of The TMX. 3) if i can run my whole test
suite (quote unquote) in 30 seconds then it's still FINE!!!!

and 4) omg i would love to visualize chokepoints somehow..