[hl] (complete overhaul of plugin) (.... 1568)
..refactor after reading `Clean Code`..) (martin)

this starts as an attempt to clean up aspects of the [cm] `config` API..
Reading (cover to cover) Martin's _Clean Code_ confirms some of our
sneaking feelings about some issues we were having. although we aren't
accepting its dogma whole-hog for everything we are doing (e.g output
arguments), we are taking a lot of its lessons to heart and trying them
out here.

  • the general trend is to go away from proxies (@plugin_parent_services) and
    towards private instance method calls
    (fulfilled with @plugin_parent_metaservices)
  • added runtime dsl to boxxy, this was the start of the whole tangent
    (in near commits previous, rewrite boxxy)
 • we re-write a really simple coverage adapter to be used for e.g
   single files
 • decide to re-write plugin from scratch using the Three Laws of TDD
• went to re-use `config` subsystem, still opaque, clunky & tightly coupled

(4 in 64.23)