[univ] protected to private in many nodes (3 of 4) (.... 1650)
hopefully this is the last huge commit like this ever.
change from `protected` to `private` in one million thousand places.
this metastasised into several refactoring concerns, all of which crashed into
each other in a predictably unexpectable way, all of which really just
stemmed from trying to DRY up the universe:

 + reduce the amount of $stdout, $stderr, etc global littering
 the place (or at least corral them into a universal convention of use),

 + reduce the number of times we create inline literals for idiomatic
 functions like -> _ { true } (better to state the idiom with a name)

DRY-ing up the "services" files in all the sub-product tier nodes lead
us to create cautiously the sub-system sub-system, which is fully
engaged now and ready to be made cautiously useful. (we are furtive
because usually when we try to make a God library we end up regretting
it somehow.)

it was "impossible" not to let sneak in overhauls to autoloading, skylab (!),
tempdir. we tried (for several days in fact) to break this in to
smaller commits but the same integration problems would come creeping
back in, for which we would have a future branch lying around in wait
that we often ended up merging. AHA! the mistake we may have made was trying
to do the protected-to-private *first* (after the big breakup), thinking it
was little and trivial, when the opposite was true.

• also we began a formal analysis of the "choke point" thing, to be
explained later..

(app tier:) tmx, permute, tan-man
(lib tier:) treetop-tools, code-molester, test-support
(framework tier:) face, headless, porcelain
(foundation tier:) basic, meta-hell

(the final integration of subsytem)

(4 in 63 sec)
(4 in 122 sec)