([f2] 2 of 4: historically green patch from during earthquake..)
this commit is kept for posterity, it was a locally green commit over the at-
-the-time rewritten headless, the rewrite that was never finished and "lost
in the fire." what follows this parenthetical paragraph is the original commit
msg. the following 2 commits after this are the re-re-written [fa] that works
on the shinest, newest headless we could muster, what with the newest API API
to boot.

the rest of this message is the original commit message:

[fl] -> [hl] (#m 3 of ~15) (~(11))