[sl]<-[*] UPGRADED to MRI 2.2.0p0 (.... 1919)
(this is the 5th of a five-part commit there.
in more detail:)

(7.24 3/5 5/5 [f2] +1) <- means..
of 7 of the 8 commits related to the git viz "dot"
milestone, this is the 24th commit in that 7th step
(yes we under-estimated how much work we would indulge
for that one). these last few commits were wedged
in as a continuation of step 3 of 5, which was the
zsh files to make fixtures (a prototype for a 2-way
system mock recorder.) (the upcoming step 4 may be
a sequence of its own, related to making a fixtures
server). (5 is cleanup & integration, and final
implementation of sibling things, and unforeseen
features/refactors). finally, the step "5 of 5"
above refers to the five steps we took upgrade to
MRI 2.2.0p0. POP!

as for [f2]:
• the change in the ancestor chain - it was always
  a dirty hack we were doing, and now it bites back
• look at that speedup!

(4 in 26.49)

• also we wedged a fire-era commit in to this one,
because we had been meaning to do something with
the patch and we wanted to keep the historical
point of that. how exciting, there's only like 5
left to rescue yay! here's the original msg:

[yacc] -> [hl] (#m X of ~15) (~(5))

  • incidentals (three lines of change) exemplifying method cherry
    picking instead of adding to chain