([hl] rework CLI box DSL..)
this is another breakout commit focused on one component, as part of
the larger plan (which seems to be the only way we would ever get
through this, is to find ways to break up huge globs of half finished
rewrites (touching sometimes 100+ files) into small pieces that still
integrate 100% with the rest of the system. this seems to be the goal
here is to master this workflow)

• the box DSL node got the usal treatment: we took long comments out of it,
brought the variable names and method names up to convention (including ugly
method names, i'm sorry the future, but they are so much nicer to work with)
and especially we put a lot of care into reworking the narrative so it
has optimal flow, including breaking long methods down and even making
helper "agent" classes.

more comments in the next commit..

([hl] doc: add to [#040] CLI box DSL narrative)