([hl] 2 of 10 rework CLI action node.. (( 121 )) )
note this is the first time we've had to fragment during this iteration.
we spent the night tearing out commits from a huge monster of changes and
committing them green against the whole system. this is the first time we
are narrowing our greenspace (and not yet integrating with the rest of the
system). part of our justification in this is that those remaining CLI
components that are not part of this commit are getting fully gutted in the
next. gutted.

the degree to which integration in the next commit or so does not incur
much change to the files from this commit will be one indication if this
fragmentation is justified. but really, who cares? it's ok to break many
eggs on the way to making the world's perfect omelete

• note large methods being replaced
• note agents breaking out from components
• note orphans getting de-orphanized (merged to parent)
• note no more goofy comments at the 35th column any more (farewell)
• note one test per "facility"