([ba][fa][hl][mh][po] normalize node references..)
we have standarized the format for node references such that a node reference
of the form "[#ba-123]" is a "foreign" node reference (rather than just a
"qualified" one). this form is now strictly reserved for making references to
that node from outside that subsystem. when we are referring to a node that is
within the same subsystem (which should usually be the case by design) we must
use the "local" form of e.g "[#123]". this allows them to be machine
searchable in a more straightforward manner.

(.... 1767)
(4 in 47.48)