([ts] 1 of 6 (moves only..))
this is the fist of six commits of a heavily planned overhaul to doc-test.
the six commits will consist of one move pair, then one huge payload
commit, then a move triad, the last of which is special.

(the move triad is a triad because we have to to move some files over to
make room for other files that need to move there to make things look
clean..)

hopefully by the sixth commit the narrative will feel balanced and whole.

the preceding clean-looking topic commits leading up to this one were all in
collusion with this madness.

we are merely freshening doc-test to add one small feature that will be
evinced by the fifth commit to have been a dubious amount of effort to this
end. WHAT FUN

• with what will become a common pattern, rename docs to "narrative" as
  appropriate.
• correct the visibility of this node
• we will integrate these changes in a separate commit from our payload
  changes because of the intensity of those changes.