[ts]<-[*] 6 of 6 doc test is more awesome now (.... 1851)
this concludes the topic chain. here is a description of what happened:
we added the ability to parse certain hashtags in our manifest of doc-test-
friendly files ('test/doc-test/data/files'). with those hashtags we can
specify options that affect how that particular file is converted to a
spec file.

specifically of those files that are "sore thumb" stem files, we want to bump
them down into their dedicated folders so they can be next to their siblings
so it is easier for us to run all specs for one node merely by specifying
one path rather than having to specify two.

sadly this is more complicated than it may sound: for one thing, if those
"stem" files once set up their own test module node, they must no longer do
so if they are to share a test module node. for another thing many of these
files are generated which means we must alter existing or make a new
template. we have done this with this commit.

to determine which template to use for the particular file we implement this
with template options. template options can be specified on the command line
but for day-to-day use it is easier to use specify them in a manifest file.

for fun we make these look like hashtags, also because it may come in handy
for some of the things we will be up to in the future. the way we implemented
this is spare and novel: the hash tag stem (not '#', not ":value") isomorphs
into a module name. we have a module of proc-shaped-modules ("bundles"), one
per "magic" hash-tag. (two currently exist, and if you use unrecognized
hashtags in your manifest file, a notice is issued and the line is processed
"normally", which will mean it is skipped because of the unrecognized data on
the line.)

anyway, that was a night well spent.

(4 in 51.92)