[*] smalls: improve deep copy implementations (.... 1851)
once we discovered that there existed an `initialize_copy` method we jumped
at it without realizing how it is meant to be used. this commmit corrects
that and upgrades our old "base_args / base_init" convention to be a bit more
mature. see [#mh-021] the dupe narrative for gory details.

distill narratives out of:
 • [hl] parameter
 • [mh] fun parse curry

(4 in 51.46)