[gv] 7.4 of 8 INTRODUCING MOCK FS (.... 1891)
please enjoy the long narrative. with our little first slash at "mock FS",
you write a list of pathnames in a manfest file. and then you can create
Pathname-looking objects from this "mock FS" that respond as expected from
things like:

 • `directory?`
 • `exist?`
 • `join`

yes it's hilarious how deep in we are in tangents. it looks like this (start
from the bottom):

     • we did a ball of mud for a [gv] revamp, and then decided to go back
       and breakout narrative chunks of it
    • the bleeding rewrite has one spot that would be perfect for our ancient
      unfinished wips of an idea, "git viz"
    • [ ??? ] something about treemap, which leads to a bleeding rewrite.
   • write so much documentation about our name conventions we decide we
     need to enlist [sg] to help process it
  • big new code near NLP, queries, and "self-describing data structures"
 • overhaul tan-man to help visualize this effort
• remove sub-client (the bad one) from the whole system (this was the work
  lost in the "fire")

(4 in 59.87)