([gv] 7.25 add "build-most-recent-rbx" task)
4/5 1/N "troika/server" pre-commit

this current commit is largely about a separate
wispy vaporware notion we are calling #autobleed.
some of the things we do here are so full of sin
that they should be printed out on a dot-matrix
printer, doused in nail-polish remover and
ceremonially cremated (hint: we use 'ed' in a
script to edit the script that is running).

the irony is this doesn't even actually build
and install rbx for you, it just prints out the
instructions to do so. it's so we don't forget
the bit about building it with openssl support.

also it has the tremendous novelty of intelligently
looking for whether any next stable version of
rbi has been published (to that one web page)
yet.

it's dirty: it uses separately curl and wget
called via open3, and doesn't use Net::HTTP
or whatever (we did it for the options).

also there is something slightly off about the
logic, that we are leaving alone for now. (it
should be checking for the existence of a
"2.2.4" on the cloud currently. it is not).

but this is the main thing: we are toying with
having autobleed work with brew so this is just
playing around with that idea; a snapshot of a
sketch if we may..

oh, also one more thing: this thing is our first
stab at a multithreaded guy but we will soon hate
whatever we did here

a quick sketch of insanity (a tangent stack,
start at the bottom):

• [this is the level for the manifest API working]
 • [this is the level for getting the server working]
        • [this level is for getting DCell working EDIT: skipped]
           • reinstall zeromq thru reinstalled brew, basic zeromq examples pass
           • rebuild rubinius now with openssl support
           • rebuild MRI (but now 2.1.0) now with our new openssl support
             (we always have trouble finding this,  brew --prefix openssl:)
             ./configure --prefix=[..] --with-openssl-dir=/usr/local/opt/openssl
             (EDIT: there is our big retarded script for this now)
           • install latest xcode before xcode 5, realize or OS is out of
             date, stop there. our startup disk is almost full..
          • resinstall brew, copy over old mvim command line executable,
            brew complained about our older version of git without ssl support,
            so with brew reinstall ack, git, openssl, wget, tree
         • when trying to install reddis we got frustrated at our brew
           installation being at odds with our macports installation,
           so we blew away /usr/local and /opt/local to start over
        • see how far we can jump into DCell
       • 3 1/2 hrs later (and after 2 hrs of dancing)
         rubinus 2.2.3 is installed and "working"
         (although we don't dare running tmx against
         it yet); and we are now reading all of the
         docs for celluloid, and about to start the
         tutuorials
      • in order to run celluloid to its full
        useful potential, we needed threads. for
        this we decided it's time to take the plunge
        (finally) into rubinius. (then we discovered
        rubinius-x, but that's crazy)
     • once we discovered that "cool.io" sort of
       morphed into celluloid.io, and this same
       guy was behind it, we got excited.
    • then when googling around for this
      (specifically "event machine alternatives")
      because we want to have at least seen all of
      our reasonably forseeable bases up front if
      not covered them. in doing so we found things
      like cool.io which we though sounded cool,
      then watched that british guy's thing about
      node.js vs em., then we read the slides
      talkin bout "don't use event machine for new
      projs"
   • googing around we read raggi's response to a
     thing about BrB being trivial to implement on
     top of event machine
  • at first we though "use DRb" like we did (sort
    of) before, but when reading the wikipedia
    page for DrB we saw BrB.
 • in a hidden way we would like it to be a server,
    so we are not re-parsing the same lines over
    and over again (here is the first false
    requirement)
• we want to implement the manifest API