([sl] 7.22 3/5 3/5 WORKAROUND SEGFAULT upgrade [..]) +6 (.... 1891))
ok: so we had an ".rspec" file in the root of
'tmx' that was turning on the rspec "--debug"
option. as part of our cleanup we decided to
exclude 'ruby-debug' from the gemset (reasonably).
rpsec has some crazy block that tries to load
the 'ruby-debug' when this option is on, and then
rescue a LoadError. for some reason a RuntimeError
was instead being issued, and something near this
was (maybe) causing the segfault. we move forward
for now, but may try to isolate this later..

because this problem and workaround wasn't reflected
in our code directly but was rather because of
a vicious dot file we had in our project root
alongside with a change in environment; because
of that we have gone ahead and changed the
.rspec.example file such that it now contains
exactly what our .rspec file had when this
occurred.

we are eager to get to the finish line because
it looks like there has been some speedup
(somehow, like in the G.C or something)

(4 in 28.19)