([gv] 7.44 renames per re-understanding ((72)) )
(the 'nontrivial restructuring' referred to in the previous commit was
eerily prophetic if understated. also, this commit commits [#030] a known
bug that we will explore in its own (probably next) commit. currently the
workaround for that bug (a flickering one!) is to start this server with
file listening disabled)

~ a case study in names:

so there was a progressively increasing sense of disorientation and
confusion wondering where we were in the system and where we might find
a node we were looking for. pretty much the *only* way we were finding
anything was by searching for it, rather than finding it where we would
"expect"; because the system(s) had become so formless, we had reached a
zen-like state of expecting nothing.

specifically what got us to this re-arch was that we were contemplating
trying to share code between client and server (specifically for reporting
manifest parse errors: we may want to do this on the logfiles in the
backend, we may want to report it in the client instead). so we would like
some node that we can float code up to that is as high as necessary while
being as low as possible.

that's when we realized that this whole system (of mock fixtures, but it
could be a system of whatever) could be broken up usefully into three
subsystems: 1) the long-running process that responds to queries about the
manifest (the "server") 2) the client that makes requests to this server
and somehow formats their responses (for *its* client, the build scripts,
a different but related system we are ignoring in this list),
and 3) the library that actually mocks the system by reading these manifest
entries and mocking them as responses to system calls (this was the
original "mock-system.rb" node, now one large-ish file).

this new understanding of our system (which came to us surprisingly late
in the game) led to the re-namings that form the bulk of this commit:

we take this opportunity to turn these re-namings into a case study about
names and taxonomy:

'manifest' (the script) -> "fixture client"
  • originally we were trying to avoid the names "client" and "server" because
    a) how overloaded in meaning "client" has become as a term and b)
    how old-fashioned it sounds having a client-server topography. however,
    that's exactly what it is and it had become too confusing not to
    employ these terms.

'fixtures/*' -> "fixture server/*"
  • same reasoning as above: since it is a server in the classical sense we
    should call it one. and now that the client and server will share
    some code (albeit in different processes) it will be especially important
    to keep these cognitively near but physically separate.
  • singularize the name ("fixture" not "fixtures"): because there will
    never be an entire server for one fixture, "fixture server" is never
    ambiguous (as "file server"). we apply the same reasoning to the client.

introduce the "manifest-entry-" node.
  • now that this has become a working concept that is coded around, it
    deserves its own node (although we would like "free tags" to float
    up and out of this node one day, perhaps eliminating it).

"isomorphic interface/*" -> [move down into server node]
  • rather than abstracting this upwards, we are pushing downwards so that
    like everything it is as high as necessary and as low as possible.

"manifest IO" -> 'manifest' (what a terrible name that was)
  • now this abstract base class names exactly what it models
    (before we were avoiding collision with the name 'manifest'
    which had become a god module or something)

'??' -> "fixture server/manifest"
  • now a class called manifest subclasses a class called manifest (sic).
    the subclass is inside of a more specialized node, so hopefully
    this cuteness makes sense and is mnemonic w/o being confusing