[cb] 1 of CB/PI plan 7 (rename [ps] -> [cb] ) (…. 1954)
The "plan 7":
7. implement "tmx git scoot" "use memory" feature
6. implement "tmx git scoot" "store memory" feature,
   using plugin arch.
5. move [gv] plugin to [hl] plugin
4. move [hl] plugin to [fa] plugin
3. move [gv] callback tree to [cb]
2. make room in [cb] for 3.
1. rename "pub-sub" to "callback"

The reason for the "plan 7":

for a while now (ok only 3 days) we had wanted to
modernify and merge some of our different event-
and callback- related efforts; and with our new
simplified "callback tree" thing from [gv], we are
very happy with it and eager to replace progressively
some "pub-sub" (now [cb] digraph) models as appropriate
to use the new simplified form or whatever. but like
all of tmx this is a tangent stack whose source
inspiration is point 7 above.

In this commit:

• that was about as difficult as we had hoped
• note that [tr] and [ta] are offline, so they
  will have catch up later (but we still had to
  keep tmx pinging ability so some refactoring
  was necessary, which we will have to undo later
  when we merge)

(4 in 26.71)