([cb] variegate 'emit' 1 of 3 (moves only))
Plan 3:

receivers and senders ("listeners" and "emitters",
"callbacks" and "hosts", etc) should know the
callback pattern that is being employed from both
sides, because of the impact it has on both the
semantics of the method calls and potentially the
logic flow. hence 'emit' has become too vague to
be meaningful.

in a 2-step, restructure [cb] "callback" (formerly
[ps] "pub-sub") so that we no longer rely heavily
on the now detrimentally ambiguous term "emit" -
with all the different kinds of emissions going on
it has become too vague to have such prominence in
the code as a thing.

1) (this commit) moves only
2) the changes necessary to get [mh] green, but
   nothing else (we will have lots more to say
   about this step in that commit)
3) the changes in the universe that follow from
   step 2