[br] <- [ts] [mh] [st] CONTOURED FIELDS BECOME ..(.. 2394)
this is significant because it is dissolving a very
old and useful library into a very new and useful
one.

completely re-write and re-conceive what was once
the "countered fields" of [mh] into the new
synthesis of the [br] "properties stack",
specifically this facility is re-conceived as
the "common frame" base-class for frames in the
properties stack (experimentally).

this "common frame" of a "properties stack" stands
in its purpose and complexity somewhere between
an actor and a model entity.

half of the builtin meta-properties and macros are
geared towards creating config structures - they
provide syntax for creating memorized prods and
methods.

the other half of the macros provide a facility
that amounts to actors with required fields, and a
'reader' flag - a bit outside the domain of the
above concerns, but historical and still useful

• [br]
  • experimentally we introduce the
    "mutable iambic scanner" - we didn't like
    mutating the iambics as we read them, but
    we employ this technique here to re-write them
    to allow for experimentation with true
    "macros" in our syntaxes
  • 'scan' is now a bit of a reserved name for
    [#cb-044] scans proper. so we replace it with
    the more general "scanner".
  • the "common frame" is a clean ground-up
    three-rules-compliant rewrite of the predecessor.
    a few scant lines from the original doc-tests
    remain (maybe). it is functionally similar although
    grammar may have changed, and near "processors"
    (formerly "absorbers") there is less complexity,
    because names for iambic readers of various
    shapes are now formalized.
  • this facility can no longer work with the
    old 

• [hl]
  • because of upstream changes, system defaults
    and system client are shorter and cleaner now yay

• [mh] while we are at it, remove old redundancy
  of "method added mixer" to [br]

(2 in 8.37)