([tm] change actor to action (.. 2434))
as it turns out, we oversimplified. it was a mistake
to make the lines scanner getter be a class method.
this task looses too much context by not having
a kernel to access. things like filesystem path
locations should not be left for class methods
floating out alone in space to determine.

if we think the API is clunky & cumbersome (which
we don't, exactly, yet) then we should fix it to
make it smooth like greased legos; rather than
avoiding using it.

this choice seals our fate that we will need some
kind of 'internal' categorization for API actions,
because it will be awful exposing this action to
other modalities - basically as soon as an action
takes in as an argument or results as a result a
first class object, all bets are off for
isomorphicizing it to other modalities in a
straightforward way.

[br]

experimentally the kernel can "reach out" to the
API so that actions don't have to awfully use the
singleton.

(2 in 8.37)