([hl]<--[fm] (2 of 2) begin to unify `find` (.. 2515))
move *ancient* find command into "place". we will have
completely plowed over the code in here once we are
finished, but as is our style we are putting this oldest
node in this new core.rb location so that our DNA is
always with us.

formerly [#sl-118] and now [#hl-171] this is one of the
oldest outstanding [#bs-017] dreams of unification.

discussion:

in keeping with the new focus of "gravity" around important
scopes, "find" is decidedly a system command and so should
live under the system node in [hl]. we put it under the
"filesystem" sub-node of this node both for "semantic
aesthetics" (i.e it just seems right) and because although
`find` is a utility and not part of the filesystem per se,
it (presumably) relies on the underlying filesystem quite a
bit to do the job it is so good at.

(2 in 8.63)