[br] `cfg.write`, contempify eventing (.. 2522)
in support of future topic, we expose a greatly
simplified reduction of the byzantine API for
writing config files: now you just pass two
strings for the section and subsection
and a scanner of pairs for the values. (oh and
an IO to write to).

with this form you don't get informational events
and all other events will be attempted to be raised
as exceptions.

along the way we turn the "git config mutable" node
to be selective-listener-aware so that we are not
needlessly creating event objects that go nowhere
in use-cases like these (perhaps especially
egregious because each inline event needs its own
singleton class).

(2 in 8.69)