[bs] search and replace `counts` (.. 2528)
The most "interesting" part of this commit is the
spec, which annoyingly gets to write to and read
from the stdin and stderr streams of a sub-process,
(which itself spawns one sub-process for `find`
and one sub-process for each chunk of paths that
we pass to `grep`). it's all terrific fun.

but this all means that with the scant two tests
that all of [bs] has, they take all of around
.3 to .5 seconds to run here. but meh hopefully
it's worth it, because for a change it feels like
we are making something genuinely useful to
ourselves.

and yeah this is eerily reminiscent of testing
interactivity on the browser.

• sadly what we though was an elegant simple
  marshaling strategy for our config files
  is (apparently) borking on strings with the
  literal two-character string "\b". #open [#020]

• (in zerk, we make our first branch node that is
   a child node, and it works pleasingly well)

• (we hack out a "toggle control")

(2 in 9.38, 8.61)