[br] beginnings of zerk API API. (.. 2559)
we add and cover a single helper class that allows
us to call a brazen app with an API call (a stream
of iambic tokens) as opposed to via interacting
with it in an interactive session on the command
line. this is kind of amazing when you think about
it, that this works at all without creating to much
blood (but yes the application has to be written
with this in mind.)

things *may* have been prettier if we had thought
of making zerk API-friendly from the start, but this
is the nature of avoiding early abstraction. as it
is it is OK. also, in an alternate dream universe
zero would merge with brazen models etc.

• "node" is a far better name than "agent" for these,
   given how tree-obsessed zerk is and how much
   irrelevant meaning "agent" already carries.

(2 in 9.71)