([bs] ZERK NON-INTERACTIVE! (.. 2568))
Both for ease of testing and what is proving to
be a fortuitous exercise unto itself, we have bent
our frontier 'zerk' application around to support
non-interactive "API" calls as well as the
simple interactive mode it was designed for.

not only do the four-ish existing actionpoints now
have coverage that were not covered before, but we
have amused ourselves in realizing that we can turn
our traditional API syntax model on its head:
whereas before the pattern for API calls was
something like:

    box [box [..]] action [ name value [ name value [..]]

here we do:

    name value [ name value [..]] box [box [..]] action

name-value pairs are provided at the level of scope
at which they are relevant. this is something we wanted
to be able to do for some syntaxes but didn't really
know how to implement it elegantly, or know if it was
too messy to bother with.

we aren't putting all our chips on this because in
the big picture it isn't very significant how our
API syntaxes are structured (or maybe it will be one
day).. but anyway it's good to know it's here if
we want it

• start to try to give the CLI-interactive client
  equal footing with the non-interactive "API"
  interface.

• as needed we have to pepper all the code with this
  experimental new ability to send special scans
  signals to release their resources. #open [#cb-059]

• no more result signals. this facility did't scale
  well and wasn't necessary. instead we prefer that
  our wacky-needs buttons explicitly call up to the
  parent logic that would have been activated by
  these signals.

• eradicate the idea of "execute/executable" and
  elevate the idea of "focus" to be "can_receive_focus"
  and "receive_focus": we want to soften the idea of
  "execution" because of the fact that execution will
  look different under different modalities & will
  often require different implementation; but we don't
  want to go so far as to hard-code the names of
  modalities as for e.g a suffix to "execute" yet.

  because "focus" is already a strong idiom in this
  library (and idiom it is - it is a variation
  of the "focus" we are familiar with) and furthermore
  because the name itself presupposes interactivity,
  we simply piggy-back and build around this concept,
  and eradicate the idea of "execute" altogether.

• introduce `is_navigational` to help implement
  the non-interactive call API.

• introduce `is_interactive` as a thing in nodes.
  the fact that this is used in application code
  violates our idealistic dream, but later for that..

[hl] • round out a description of our
       state-of-the-moment name conventions for
       event-related methods and tags

     • use our newfangled and still cooling "signals"
       API to make scans (`grep` and `find`) that can
       terminate the sub-processes before they reach
       their own "natural" exits. Now we can release
       these resources early in our "long running"
       processes.

(2 in 10.15)