[br] SEARCH AND REPLACE IS KEWEL. (.. 2571)
With one final little lump of mud (one large-ish
procedural, non-"streaming" thing), we finish off
this last major desired feature from our wishlist.

we were not able to re-use as much as of the non-
interactive API code for this as we had hoped but
meh.

now all there is left to do is use it (and that
bug with marshaling escape sequences (and tons of
uncovered edge cases (but meh this could have just
been a one-off-esque (maybe))))

with our exemplary and simple use-case committed
inline with this commit, we made the changes in a
few seconds with a minimal number of interactions.

we did hit one issue that we are postponing: just
one of the matches didn't change ([bz]/core.rb).

because we can, here is the transformation:

[curried-search-and-replace-agent "current"]
dirs = .
files = *.rb *.md
replace = ACHIEVED
search = "/\\bACHEIVED/"

[curried-search-and-replace-agent "current"]
grep-via = grep

(2 in 9.75)