([cm][cs][cu][de][sla][sli][fl][fm][gi][ttt] :✓+[#libricate] (.. 2571))
With a little manual intervention necessary this
procedural refactor went off mostly as planned.

sometimes we bump our heads on the ceiling of the
limitations of this being ultimately a very
overwrought and thoroughly engineered regex hack.
(knowing when parenthesis are necessary for e.g is
non-trivial).

it would be “nice” to have language-aware refactoring
tools.. but we are waiting till the next language for
that.

sorry for potato, what follows is the gist of the
two transformations we are using. enjoyably we straddle
making the regex compatible with both oniguruma
and grep so in one case (?) we don’t escape the
‘]’ in the character class. also we apparently don’t
have lookbehind in grep, so we hack it again.

    search = "/([^:a-zA-Z])Lib_::([A-Z][A-Za-z0-9_]+[^_])\\[([^\\[]*)\\]/"
    replace = {{ $1 }}LIB_.{{ $2.transform_name }}{{ $3.args_part }}

    search = "/::Lib_::([A-Z][A-Za-z0-9_]+[^_])\\[([^]]*)\\]/"
    replace = ._lib.{{ $1.transform_name }}{{ $2.args_part }}

    rx = %r(\A[A-Z](?=[a-z0-9]+))

    camel_class = %r(\A(?:[A-Z][a-z0-9]+){2,}\z)

    Transform_name = -> string do

      if camel_class =~ string
        name = ::Skylab::Callback::Name.via_const string.intern
        name.as_lowercase_with_underscores_symbol.id2name
      else
        string.gsub rx do
          $~[ 0 ].downcase
        end
      end
    end

    Args_part = -> string do

      s = string.strip

      if s.length.zero?
        s
      else
        " #{ s }"
      end
    end

all in all, kinda great. what we had hoped for. the
power to go ham cray when necessary.

(2 in 9.43)