([bs] feature-add to S & R: escape sequences in replacement )
The feature added here will be used for a changeset in the future
but since we put S & R to a big huge use-case in the next changeset
we just folded this feature in now instead of messing around with
keeping the narrative solid.

The feature added here is that in the replacelement expression you
can now indicate newline, tab (and backslash) with the usual
escape codes.

these escape/unescape pairs apply correctly to incoming user input;
they survive the roundtrip through marshaling and unmarshaling,
and they apply correctly to strings before they are outputted as
UI.

[br] CLOSE [#066] cool new buildless events